+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.reactnavigationdemo",
+    package = "com.weipay",
 )
 
 android_resource(
     name = "res",
-    package = "com.reactnavigationdemo",
+    package = "com.weipay",
     res = "src/main/res",
 )
