+++ android/app/src/main/java/com/weipay/MainActivity.java
@@ -1,4 +1,4 @@
-package com.reactnavigationdemo;
+package com.weipay;
 
 import com.facebook.react.ReactActivity;
 
@@ -10,6 +10,6 @@ public class MainActivity extends ReactActivity {
      */
     @Override
     protected String getMainComponentName() {
-        return "reactNavigationDemo";
+        return "WeiPay";
     }
 }
