+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "reactNavigationDemo",
-  "displayName": "reactNavigationDemo"
+  "name": "WeiPay",
+  "displayName": "WeiPay"
 }
