+++ ios/WeiPay-tvOS/Info.plist
