+++ ios/WeiPay/Images.xcassets/AppIcon.appiconset/Contents.json
