+++ ios/WeiPay/Images.xcassets/Contents.json
