+++ ios/WeiPayTests/Info.plist
