+++ ios/WeiPayTests/WeiPayTests.m
@@ -14,11 +14,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface reactNavigationDemoTests : XCTestCase
+@interface WeiPayTests : XCTestCase
 
 @end
 
-@implementation reactNavigationDemoTests
+@implementation WeiPayTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
