+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "reactNavigationDemo",
+  "name": "WeiPay",
   "version": "0.0.1",
   "private": true,
   "scripts": {
